refactor: test, types
