+++ jest.config.js
@@ -3,6 +3,6 @@ module.exports = {
     transform: {
         '^.+\\.ts$': 'ts-jest',
     },
-    roots: ['<rootDir>/src'],
+    roots: ['<rootDir>/tests'],
     testMatch: ['**/*.test.ts'],
 }
