+++ package.json
@@ -30,9 +30,9 @@
         "prepack": "yarn build",
         "prebuild": "rm -rf dist",
         "build": "tsc --project .",
-        "lint": "eslint src/**/*.ts",
+        "lint": "eslint tests/**/*.ts src/**/*.ts",
         "lint:fix": "yarn lint --fix",
-        "test": "jest src",
+        "test": "jest tests",
         "test:watch": "yarn test --watchAll",
         "test:coverage": "yarn test --coverage",
         "types": "tsc --noEmit src/**/*.ts",
