+++ tests/utils.test.ts
@@ -1,4 +1,4 @@
-import { convertSizeToBytes } from '../utils'
+import { convertSizeToBytes } from '../src/utils'
 
 describe('utils', () => {
     it.each`
