+++ tsconfig.json
@@ -69,5 +69,5 @@
     "skipLibCheck": true,                           /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
     },
-    "exclude": ["dist/**/*", "src/__tests__/**/*"]
+    "exclude": ["dist/**/*", "tests/**/*"]
 }
